Remove the use of gotestfmt for pulumi-kubernetes
Using `gotestfmt` appears to be more trouble than it's worth -- see
https://github.com/pulumi/pulumi-kubernetes/actions/runs/3490648026/jobs/5842448567
for example, where there are a number of apparent failures for tests
that actually passed (note the X by "TestRetry", then search for
"PASS: TestRetry").

This is a conservative change to remove its use just for the
pulumi-kubernetes provider. If it's causing problems elsewhere too, we
can make a more sweeping change.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>