Fix to allow safe automerge
Fixes #272

Previously, automergable PRs had automerge enabled externally immediately after the PR was created
assuming that only if the sentinel check passed, the PR would work. However, in case of failure in tests
the sentinel check is skipped entirely causing the PR to be immediately merged when we don't want.

This change tries to enable automerge from within the sentinel step which will execute when the steps
it depends on are successful.