Rebuild templating using Go text templates
Seeking to resolve the CI disparity between ecosystem managed repositories and
pulumi/pulumi-std#29, it quickly became apparent that the style of GitHub
Actions - using reusable workflows - in that repository and in pulumi-command
would require a major rewrite of the templating in this repository.

Making large changes to the templating in this repository is more difficult as
it requires modifying the Node scripts and the type-safe "generator" to reverse
engineer CI workflows into a known form.

Directly generating text templates, we skip the intermediate steps and can more
naturally write the intended GitHub Actions YAML.

The Go utility in this repository can be run and a diff with the current
repository scripts generated via:

```
cd ./package-ci
./scripts/test.sh
```

The committed template, "bridged-provider", is intended to be bit-for-bit
identical in output to the current scripts to the extent possible. As a result,
there is no semantic diff with the current output.

The two cases where the output was not identical are in the New Relic provider,
and the diff is inconsequential wrapping behavior in the current Node scripts
that is not present in the Go text template, both appearing in `.goreleaser`
files. In this case, the wrapping has no effect on Goreleaser:

```diff
diff --git a/../provider-ci/providers/newrelic/repo/.goreleaser.prerelease.yml b/./test-output/providers/newrelic/repo/.goreleaser.prerelease.yml
index 439adbf7a..a5b009761 100644
--- a/../provider-ci/providers/newrelic/repo/.goreleaser.prerelease.yml
+++ b/./test-output/providers/newrelic/repo/.goreleaser.prerelease.yml
@@ -29,8 +29,7 @@ builds:
   ignore: []
   ldflags:
   - -X github.com/pulumi/pulumi-newrelic/provider/v5/pkg/version.Version={{.Tag}}
-  - -X
-    github.com/newrelic/terraform-provider-newrelic/v2/main.UserAgentServiceName=pulumi
+  - -X github.com/newrelic/terraform-provider-newrelic/v2/main.UserAgentServiceName=pulumi
   main: ./cmd/pulumi-resource-newrelic/
```