Add an option to clean go cache between GoReleaser builds
This chanage addresses a blocker in pulumi-azure.

The cross-build job started failing inside GoReleaser with out-of-disk-space error.

By cleaning Go cache between every architecture build, more wall-clock time is spent but critically less peak disk space
is used so the job can succeed again.

This flag assumes that GoReleaser is called with `-p 1` so it does the cross-builds sequentially, which seems to be the
case for pulumi-azure.