Preserve commit metadata on rebasing patches
Modifies the `make upstream.rebase` and `make upstream.finalize` targets in repo
Makefiles to use `git am` to "apply a series of patches from a mailbox", instead
of `git apply`. From `git am --help`:

> Splits mail messages in a mailbox into commit log message, authorship
> information and patches, and applies them to the current branch.

This git subcommand is designed to use `git format-patch` output, preserving the
metadata from the original patch, unlike `git apply` which applies a diff to the
current working tree. As this command replaces both `apply` and `commit`, we
also remove the `git commit` command from the loop.

The `make upstream` target is unchanged. That target, which invokes the
`apply()` function in `scripts/upstream.sh`, still uses `git apply`. That
command is intended to leave the user with a dirty submodule, where the ref of
the submodule points at a real upstream commit hash and the patches are staged
changes. That makes it possible to stage and commit the submodule (which only
updates the ref it points to) and have it be navigable/usable by other users.