Make `sentinel` useful when determining workflow success (#544)
* Make `sentinel` useful when determining workflow success

Currently, each repository must set branch protection for a host of jobs to prevent bad
PRs merging (for example: https://github.com/pulumi/pulumi-kafka/pull/200). This per
repository configuration is exacerbated when the set of ci-mgmt generated jobs fails:

1. If a job is added, then the repo is not actually protected from that job failing until
branch protection is manually updated.

2. If a job is removed (and it was required before), no PRs will merge since GitHub waits
for the job to succeed before allowing the PR to merge, and the job will never run.

The solution is to provide an abstract required status check, and have it determine if a
PR is safe to merge. `sentinel` is the obvious job to choose. This PR will allow
repositories to set `sentinel` as a required status check and get expected behavior.

* Update 1 provider: kafka

* Update other providers