Simplify the lint job (#576)
Fixes https://github.com/pulumi/ci-mgmt/issues/574

I think the root cause is some version mismatch between Go tooling and
golint version. I found the job really strange though so took the time
to make it more in line with the docs on golangci-lint itself and how we
lint the bridge.

Use official golangci/golangci-lint-action@v3 action

Pro: caches
Con: drift from Makefile to CI

Do not build the provider in the lint job.

Pro: faster, fewer things to install
Con: need to disarm go:embed directives; the linter checks if the
embedded files are not present, which they will not be before a lot of
computation

Do not containerize the lint job.

Pro: use same ubuntu-latest runtime environment as the rest of the jobs.