Fix missing deps for sourcemigrator (#610)
Fixes https://github.com/pulumi/ci-mgmt/issues/609

Here's what happened. As part of rolling out wheels to providers, we've
coded automatic upgrades as part of the source-migrator steps in
TypeScript, and these steps assume additional build dependencies:

- python
- pulumictl

When testing the workflows on gcp,aws,tls,random, the dependencies were
added to some but not all of the workflows. That's he omission currently
corrected by this PR.

When wheel rollout is complete we remove both the migrations and the
dependencies. Separately we might also consider removing
update-workflows-ecosystem-providers since the team split is no more and
we're now operating under one risk profile/tradeoffs not two.