Template PULUMI_CONVERT rollout (#655)
To be reviewed with https://github.com/pulumi/pulumi-azuread/pull/484

Part of https://github.com/pulumi/pulumi-terraform-bridge/issues/1280

As a pre-requisite to PULUMI_CONVERT being enabled, it's necessary to
pin Pulumi CLI so that examples continue to be deterministic. This PR
borrows targets from pulumi-azure-native that perform CLI pinning.

## Updates

### CLI 

Unlike azure-native, it assumes that the CLI version we want to use to
convert examples is the same as pulumi/pkg version we want to use to
build this provider. This removes a degree of freedom but buys automatic
upgrades - the existing mechanisms that update pulumi/pkg (in
upgrade-provider) should also update this Pulumi CLI version.

### Converter

This is not enforced here but
https://github.com/pulumi/pulumi-azuread/pull/484 adds TF converter
pinning to the CI-mgmt plugins section. This PR makes changes to allow
opting into converter and not just resource plugins. There is currently
no mechanism to upgrade this, but presumably upgrade-provider can build
mechanisms to upgrade all plugins referenced in this config file. Any
such upgrade is likely to trigger the need to regenerate SDKs.