This adds a license check on PRs to providers
This is necessary to avoid someone manually reviewing each provider upgrade for license
compliance. This PR as introduced is a required check.

pulumi/license-check-action is not total, so we also maintain an allow list of paths that
are not checked. I have assumed that golang.org and github.com/golang will only contain
repos that are valid to include, since they are both open source projects. We own
github.com/pulumi, so I have marked that safe as well (it was flagged as unable to find
license on the bridge). Other whitelisted repos I have manually checked for a OSS
License (Apache, MIT or BSD).