Enable setting custom Pulumi CLI version to install for native providers (#714)
This PR changes the Pulumi CLI installation step to use
`pulumi/actions@v4` which allows us to pin a specific version of the CLI
to install. A new config variable: `pulumiCLIVersion` allows us to
specify if a version should be pinned.

This PR also pins the Pulumi CLI version to v3.89.0 for the Kubernetes
provider.