Add devbox config to providers (#713)
Adds templates for devbox.json and .devcontainer to bridged providers
with the goal of simplifying environment setup for building providers.

As part of this work, I refactored the gradle version to be explicit in
the config file so we can ensure that the devbox environment uses the
same version.

I'm explicitly not including the pulumi cli or any plugins in the devbox
environment. Keeping pulumi up to date in nixpkgs is not something I
want to take on right now, and we already have a decent system for
loading a pinned version in the Makefile (at least for the bridged
providers).