Update Makefile detection of java version (#747)
As part of https://github.com/pulumi/ci-mgmt/issues/506, we are setting
the Java version in a provider-level version file, acting as our
"lockfile".

This PR mainly updates the provider Makefiles to be in sync with the
changes propagated with the latest (and hopefully last) bulk Java
upgrade. [Sample update
here](https://github.com/pulumi/pulumi-civo/pull/348). The rollout has
been a bit messy; there are currently failing updates in a few
providers. They should clear up as soon as this change is merged.