Fix setup pulumi pin (#755)
Noticing that AWS was testing against outdated CLI v3.77.1. This is
unfortunate.

This change gets it back to floating as all the other providers. All the
other providers are updated to use the latest recommended GitHub action.

I have slight preference for pinned and weekly updating but that's more
setup, for now floating is better than pinned indefinitely in the past.