Do not check upstream updates for pulumi-mysql (#768)
Due to https://github.com/pulumi/pulumi-mysql/issues/145 there are
currently no actionable updates on this provider. We should skip the
update check, as it fails due to an archived repo.

The main effect of this change is that it will pacify upstream check
failures.

Fixes https://github.com/pulumi/pulumi-mysql/issues/286.