Run acceptance tests on all branch targets (#785)
I found it surprising when trying to [stack](https://stacking.dev) some
PRs that we don't run tests unless the PR is explicitly against master.
This is
[awkward](https://github.com/pulumi/pulumi-docker/pull/948#issuecomment-1897531359)
to work with.

This PR changes our workflow to trigger for PRs against any branch. I
might be missing some context around #723 and why we chose to opt-in
only some patterns, but if there are automated flows that shouldn't run
tests I recommend we opt those out explicitly.