Allow running the acceptance tests in pulumi/examples as part of the test suite (#823)
This change adds a new optional flag `testPulumiExamples` that can be
set in .ci-mgmt.yaml. When set, the acceptance tests in pulumi/examples
will be run as part of the test suite.

The goal is to increase test coverage in provider PRs or releases.
`pulumi/examples` has a rich set of realistic programs that can be used
for this via ProgramTest.

Providers don't necessarily need to gate PRs on these tests, they can be
informational only.

This requires that the provider's test suite in `examples/` is enabled
for this. See pulumi/pulumi-azure#1717 for an example. It's not enough
to just run the test suite in pulumi/examples because we want to use or
inject the locally built provider and SDKs of the current PR.

I chose the approach using the test job's matrix because the new
pulumi/examples tests should run concurrently to the regular ones, but
to run them in a separate job would require copying the lengthy setup
steps of the test job, or extracting the setup into a composite action
which I thought could be avoided.