chore: create upstream patch files with `--no-numbered`
Currently the patch files are generated with subjects that contain the
patch number (i.e. `Subject: [PATCH 02/52] Add S3 legacy bucket to resources`).
This makes diffs too verbose since everytime I add or remove a patch I
have a diff on _every_ one of the patches.

The patches have the patch number in the name of the file already so the
number in the subject is redundant.