Make generate cli arguments optional (#937)
Related to https://github.com/pulumi/ci-mgmt/issues/936 - this is a
pre-cursor to building out composable templates.

## Unbundle the config parsing from the generate method

We need to parse the config before calling generate.

## Make it simpler to run the provider-ci binary directly

Make the 4 required arguments to `provider-ci` optional:

```
Usage:
  provider-ci generate [flags]

Flags:
  -c, --config string     local config file to use (default ".ci-mgmt.yaml")
  -h, --help              help for generate
  -n, --name string       repository name to generate (default "{config.repository}" or otherwise "pulumi/pulumi-{config.provider}")
  -o, --out string        directory to write generate files to (default ".")
  -t, --template string   template name to generate (default "{config.template}" or otherwise "bridged-provider")
```

This avoids the need to generate a makefile template to make it easier
to run the tool to pull templates:

```diff
- go run github.com/pulumi/ci-mgmt/provider-ci@master generate \
-		--name pulumi/pulumi-$(PACK) \
-		--out . \
-		--template bridged-provider \
-		--config $<
+ go run github.com/pulumi/ci-mgmt/provider-ci@master generate
```

Existing providers have all the required information available except
for the `template` field.