Fix `make upstream.rebase` documentation
In effect `make upstream` (`apply` in `upstream.sh`) is called as the final step in `make
upstream.finalize`. Because `make upstream` resets the upstream submodule to the currently
staged commit, if `TO` is set to any value that is not HEAD it will be ignored (but look
like it worked).