Implement new explicit sentinel check (#944)
Write an explicit status check called "Sentinel" which will only pass if
this code really runs. Once rolled out, we can make this the only
required check for PRs, then remove the old conditionals on this job and
remove the previous step.

This is a follow-up to the conversation
https://github.com/pulumi/ci-mgmt/pull/942#issuecomment-2137523033 where
we debated the merits of a single required check if we could simplify
the current sentinel logic to make it more reliable.

Here's the test run in the xyz provider:
https://github.com/pulumi/pulumi-xyz/actions/runs/9292939832?pr=145

Here's the old and new sentinel checks shown on the [test
PR](https://github.com/pulumi/pulumi-xyz/pull/145):

![image](https://github.com/pulumi/ci-mgmt/assets/331676/2a9ac907-729e-415b-aac6-87ff29eb9d7b)