Update native provider CI
The existing native providers still require the existing sentinel job name as well as the build and test steps.
This adds an additional check which we can then switch to be the only step we depend on once it's rolled out.
The existing sentinel job check had the issue where if previous jobs failed, it would be skipped which is considered complete by the branch protection.