Add option for freeing disk space (#964)
Stacked on top of #942 

- Provide dedicated option instead of relying on providers to use custom
steps which are run after tool installs.
- Run as first step so we don't delete things that have just been
installed.
- Default to off to maintain existing behaviour. This can take around 3
minutes to run, so we should only run it if we're having issues.

```yaml
# Set to true to clear disk space before running prerequisites workflow.
freeDiskSpaceBeforeBuild: false
# Set to true to clear disk space before running test jobs.
freeDiskSpaceBeforeTest: false
```

This fixes oddities in Azure and AWS where we've added steps in the
`preBuild` and `preTest` to clear disk space but then have to install
tools (e.g. dotnet) a second time as they get deleted.