Encapsulate tool setup into a composite action (#965)
Stacked on top of https://github.com/pulumi/ci-mgmt/pull/942

Primary benefit is to reduce duplication within templates. This was not
possible when using splices because splices cannot reference splices.

Additionally, we want to make installing our standard versions of our
tools available for use from custom test jobs as used in AWS and other
providers so they don't have to hard code their own install actions.

- Go was not previously explicitly installed during the build_sdk job
but is used when running `make build_go`, so this adds it to ensure
we're using a consistent version.