Only calculate version once in prerequisites (#978)
This ensures that the version is always consistent between different
jobs within the workflow even if external dependencies change (e.g. a
new release goes out half way through).

This is also required for us to be able to call a new re-usable workflow
which requires the version to be available as a job output.

Tidy up the job ordering so prerequisites is the first job in the
workflow (as it's the first to run).