Verify release action (#977)
Add new job in our release workflows to verify that our released SDKs
and provider are usable.

This will have to be configured into each provider as we will need a
program specified in the local repo for each runtime we want to verify.
Ideally this will include: nodejs, python, dotnet and Go. Without
configuration, the job will not do any verification. There's an example
of the available configuration as documentation in the
`bridged-provider.config.yaml`.

The verify-release workflow is also runnable directly if we want to do a
one-off verification. See the
[pulumi/verify-provider-release](https://github.com/pulumi/verify-provider-release)
action for the underlying implementation.

Fixes #972 

[Example run without
configuration](https://github.com/pulumi/pulumi-xyz/actions/runs/9604446619)
skips the job:


![image](https://github.com/pulumi/ci-mgmt/assets/331676/a14dca95-5450-4ccc-b5f2-d2ff3cc9f276)

[Example prerelease with nodejs
configuration](https://github.com/pulumi/pulumi-xyz/actions/runs/9604468261)
runs on Ubuntu and Windows:


![image](https://github.com/pulumi/ci-mgmt/assets/331676/80b11edf-94da-4343-a0f3-163ad40a7d7f)