Use pulumi cli version from `.pulumi.version`
The providers have a `.pulumi.version` file which stores the current
pulumi version. There is also a job `weekly-pulumi-update.yml` that
updates this version.

In our CI jobs we install the pulumi cli using the `pulumi/actions`
action which uses the [version installed on `ubuntu-latest`](https://github.com/actions/runner-images/blob/c5c9a7cd7b2cfd20d740f7667fdaca8a412f2b52/images/ubuntu/Ubuntu2404-Readme.md?plain=1#L92) by default.
This means that we are not testing on the version that we want to.

This introduces an additional step which sets the pulumi version to
download based on the value from that file.

I've tested this strategy on this PR in `aws-native` and it seems to be
working.