Allow composable templates (#989)
Separate the concept of a template package from the template directories
it's going to render.

A template is formed of one or more directories. Each template directory
be rendered into the same output directory.

This forms part of https://github.com/pulumi/ci-mgmt/issues/936 where we
want to be able to share some parts of the current bridged provider
template with non-bridged provider repos.