Simplify provider-ci makefile (#993)
Rewrite example provider testing targets to be more linear and easy to
follow. This is a clean-up in order to add more providers to test as we
start to migrate new ones in (i.e. native providers & components).

- Move "test-workflows" to "test-providers" - we're generating more than
just GH workflows.
- Move config files for testing into each provider repo - named
.ci-mgmt.yaml - just like the real repo.
- Move temp linting repo into the bin directory which is fully ignored
as it's just temp files.
- This also fixes the random new fiiles showing up in the git working
tree.

This is now slightly faster for running both clean and repeated runs of
`make`.

Clean run on `make`:
- Old: `2.40s user 2.76s system 159% cpu 3.224 total`
- New: `1.39s user 1.77s system 127% cpu 2.480 total`

Repeated run of `make`:
- Old: `1.63s user 1.99s system 231% cpu 1.561 total`
- New: `0.63s user 0.95s system 243% cpu 0.649 total`