Make log outputs optional in the issue template (#1024)
We don't need log outputs if a clear repro is given for an issue, so
it's unnecessary to require that. It raises the barrier to reporting an
issue, possibly costing us bug reports.

I haven't seen a pulumi engineer add logs when reporting an issue,
except when that's meaningful, so we shouldn't expect outside
contributors to do that.