Relax unforced validation
Relax validation for running 'init' and 'checkout' without the force option. Only require upstream to not be checked out on the 'pulumi/patch-checkout' branch, but don't require it to be tracked in git. This allows the unforced init to be used after pulling changes but not having yet updated the submodule.