Provide better instructions when upstream is modified (#1045)
- Recommend adding `-f` rather than custom git checkout commands.
- Hide the `format_patches` command as `check_in` is a much less
confusing version of the command.
- Don't use the force option from the makefile as we don't want to wipe
out any uncommitted work by accident.
- Relax validation for running 'init' and 'checkout' without the force
option. Only require upstream to not be checked out on the
'pulumi/patch-checkout' branch, but don't require it to be tracked in
git. This allows the unforced init to be used after pulling changes but
not having yet updated the submodule.
- Assert we're not mid-rebase if force not provided (avoids accidentally
loosing progress if switched to our own branch).
- Clean up old branches and rebases when using the force option with
either init or checkout.

Example error (for init):
```
Error: 'upstream' submodule checked out on the 'pulumi/patch-checkout' branch.
This was likely caused by running a 'checkout' command but not running
'check_in' afterwards.

To turn the commits in the 'pulumi/patch-checkout' branch back into patches, run:
  ./upstream.sh check_in

To disgard changes in the 'pulumi/patch-checkout' branch, use the 'force' flag (-f):
  ./upstream.sh init -f
```