Remove gotestfmt (#1069)
Engineers noticing gotestfmt reporting test output in a confusing way,
such as mis-attributing logs from one test to another or not reporting
logs if a panic is encountered. This change removes this utility,
switching to default Go test output. It also removes the `-v` flag so
that the test failure output is minimally scoped to the failing tests
and is easier to find. Defaults for the win.