Ability to select SDK languages (#1063)
**Update Sep 9**: dropped the `enableAutoRelease` from this PR. Will
redo this as a separate PR.

Round 2 to add support for third-party providers:

* Fix: Pass the configured `organization` to 2 more places where it is
needed
* `languages` (default: all supported languages): third party providers
can configure a subset of the supported languages if they don't have the
publishing infrastructure ready for all the languages. Changes:
  * changes the build matrix to only the set of configured languages
* will install only the tools needed for the given value of the language
matrix. Changed this because installing the Java tools failed when no
Java SDK is available.

<img width="1621" alt="Screenshot 2024-08-28 at 15 00 07"
src="https://github.com/user-attachments/assets/eba0eeba-6d40-4e33-ba12-21775c0a2f4f">

The changes in this PR are focussed on the `main` workflow. Testing this
on `pulumiverse/pulumi-acme` already got me this far:

https://github.com/pulumiverse/pulumi-acme/actions/runs/10596199445