Add registry docs option to bridged provider CI config (#1083)
This pull request adds a `registryDocs` configuration option to the
provider CI.

When enabled, it will run `./bin/pulumi-tfgen-foo registry-docs --out
docs/`, automatically writing upstream installation documentation.

Note that in the Makefile, the functionality is added as a dependency to
`make_build_sdks`, for the following reasons:
1. In order to generate docs, the tfgen binary must be built, and the
`build` commands are based on the same assumption
1. These files are auto-generated; just like the SDKs they should be
generated at the end of an update
1. It avoids adding a new separate target all over our github Workflows
and upgrade-provider.

Part of https://github.com/pulumi/home/issues/3598.