Inline upgrade provider action (#1097)
This action was originally failing due to missing dotnet if we run the
disk clean-up action first. Then when trying to install dotnet using the
setup-tools action, it was failing because the repo wasn't yet checked
out, however, the `pulumi-upgrade-provider-action` does it's own
checkout action, so it would have required removing the checkout from
that action, creating a new version, then upgrading the version via
ci-mgmt. At this point the `pulumi-upgrade-provider-action` would only
contain the install and execution of the upgrade-provider CLI tool.

## Change summary

- Go direct to the `upgrade-provider` program instead of via the
`pulumi-upgrade-provider-action` wrapper action.
- Enables us to use the shared setup-tools action for tool consistency.

## References

- Existing wrapper action:
https://github.com/pulumi/pulumi-upgrade-provider-action/blob/main/action.yml
- Sample run:
https://github.com/pulumi/pulumi-azure/actions/runs/11219063931/job/31184185595