Run tests for thirdparty packages (#1102)
I moved `run-acceptance-tests.yml` in #1084 to the `pulumi-provider`
folder as I thought it was Pulumi Corp specific. I'm moving this back to
`bridged-provider` folder because it is also needed for third-party
provider packages. The workflow is updated to use the `GITHUB_TOKEN`
with elevated permissions for commenting on the PR.