Combine upstream check & upgrade workflows (#1110)
Remove the chain of dependencies where we create issues then trigger the
actual upgrade from the issue creation. We previously had to filter the
triggered issue opening to a specific bot user to stop other issues
created by external users from triggering the workflow.

Instead, we'll check for new versions then attempt the upgrade in a
single run, as well as allowing the execution of a specific version
upgrade via workflow_dispatch.

Scenarios:

- If the cron triggers or a user runs the workflow without a version,
it'll ensure an issue exists for the latest pending version and return
the latest pending version in the `latest_version` output.
- If `latest_version` is unset then we don't need to run the upgrade as
there's no pending version.
- If `latest_version` is set then we'll specifically attempt an upgrade
to that version.
- If a user runs the workflow with a version input, we'll skip creating
issues and just attempt the upgrade.

This builds on the work in
https://github.com/pulumi/upgrade-provider/pull/282

Fixes https://github.com/pulumi/ci-mgmt/issues/1108

Example run with no new changes:
https://github.com/pulumi/pulumi-azure/actions/runs/11608904415/job/32325098437

---------

Co-authored-by: Ian Wahbe <me@iwahbe.com>