Don't fail the upgrade check job if we can't open a PR (#1129)
Docs for continue-on-error:
> Prevents a job from failing when a step fails. Set to true to allow a
job to pass when this step fails.

Failing the job results in a P1 issue being opened. We've already opened
an issue to track the pending upgrade, so we'll just exit cleanly if the
PR can't be made.

This will still raise a P1 issue if the upstream check (to create the
tracking issue) fails.